Remove system pool entirely
Presto has been running fine without a system memory pool in our
production clusters for quite some time. This change removes it entirely.

Extracted-From: https://github.com/prestodb/presto
