+++ presto-main/src/main/java/io/prestosql/memory/NodeMemoryConfig.java
@@ -15,6 +15,7 @@ package io.prestosql.memory;
 
 import io.airlift.configuration.Config;
 import io.airlift.configuration.ConfigDescription;
+import io.airlift.configuration.DefunctConfig;
 import io.airlift.units.DataSize;
 
 import javax.validation.constraints.NotNull;
@@ -22,13 +23,13 @@ import javax.validation.constraints.NotNull;
 import static io.airlift.units.DataSize.Unit.BYTE;
 
 // This is separate from MemoryManagerConfig because it's difficult to test the default value of maxQueryMemoryPerNode
+@DefunctConfig("deprecated.legacy-system-pool-enabled")
 public class NodeMemoryConfig
 {
     public static final long AVAILABLE_HEAP_MEMORY = Runtime.getRuntime().maxMemory();
     public static final String QUERY_MAX_MEMORY_PER_NODE_CONFIG = "query.max-memory-per-node";
     public static final String QUERY_MAX_TOTAL_MEMORY_PER_NODE_CONFIG = "query.max-total-memory-per-node";
 
-    private boolean isLegacySystemPoolEnabled;
     private boolean isReservedPoolEnabled = true;
 
     private DataSize maxQueryMemoryPerNode = new DataSize(AVAILABLE_HEAP_MEMORY * 0.1, BYTE);
@@ -47,18 +48,6 @@ public class NodeMemoryConfig
     public NodeMemoryConfig setMaxQueryMemoryPerNode(DataSize maxQueryMemoryPerNode)
     {
         this.maxQueryMemoryPerNode = maxQueryMemoryPerNode;
-        return this;
-    }
-
-    public boolean isLegacySystemPoolEnabled()
-    {
-        return isLegacySystemPoolEnabled;
-    }
-
-    @Config("deprecated.legacy-system-pool-enabled")
-    public NodeMemoryConfig setLegacySystemPoolEnabled(boolean legacySystemPoolEnabled)
-    {
-        isLegacySystemPoolEnabled = legacySystemPoolEnabled;
         return this;
     }
